[AC-9028] move functions to coreprofile model
